KeyUp + Code Reload
As I was thinking of implementing the conductive patches with the makey makey and the RPi, I thought "hey, Makey Makey triggers as long as it's held, that's going to be an issue for the konami code listener". It was.

So, switched from keypress to keyup.

Also implemented a timer that resets the page every 3 minutes for hands-free/click-free reset.